backend/deployment/agentcore/requirements.txt
# AWS Bedrock AgentCore Dependencies
# Core AgentCore SDK and tools
bedrock-agentcore<=0.1.3
bedrock-agentcore-starter-toolkit

# AWS SDK for manual deployment
boto3>=1.26.0
botocore>=1.29.0

# Strands Agent Framework
strands-agents
strands-agents-tools

# Container and deployment tools
docker>=6.0.0
requests>=2.28.0

# Async support
aiofiles>=22.1.0
asyncio-mqtt>=0.11.0

# Logging and monitoring
structlog>=22.1.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
